Fixed the following warnings:  - Assignement within conditional expression.  - Unreachable code  - Potentially uninitialized local variable used  - Conversion from 'x' to 'y', signed/unsigned mismatch  - Conversion from 'x' to 'y', possible loss of data  - Function uses 'x' bytes of stack. Consider moving some data to heap.  - 'x' is allocated with array new [], but deleted with scalar delete  - Redundant code. The left and right sub-expressions are identical.  - Mutual exclusion over || is always a non-zero constant.  - Argument 'x' might be 'y'. This does not adhere to the specification for the function 'z'.

+++ src/arcemu-shared/CrashHandler.cpp
@@ -299,13 +299,13 @@ int __cdecl HandleCrash(PEXCEPTION_POINTERS pExceptPtrs)
 	/* only allow one thread to crash. */
 	if(!m_crashLock.AttemptAcquire())
 	{
-		TerminateThread(GetCurrentThread(),-1);
+		TerminateThread(GetCurrentThread(),static_cast<DWORD>(-1));
 		// not reached
 	}
 
 	if(died)
 	{
-		TerminateProcess(GetCurrentProcess(),-1);
+		TerminateProcess(GetCurrentProcess(),static_cast<UINT>(-1));
 		// not reached:P
 	}
